+++ crates/proving-service-client/src/proving_service/block_prover.rs
@@ -3,6 +3,7 @@ use alloc::{
     sync::Arc,
     vec::Vec,
 };
+use core::time::Duration;
 
 use miden_objects::{
     batch::ProvenBatch,
@@ -66,9 +67,16 @@ impl RemoteBlockProver {
 
         #[cfg(not(target_arch = "wasm32"))]
         let new_client = {
-            ApiClient::connect(self.endpoint.clone())
+            let endpoint = tonic::transport::Endpoint::try_from(self.endpoint.clone())
+                .map_err(|err| RemoteProverError::ConnectionFailed(err.into()))?
+                .timeout(Duration::from_millis(10000));
+            let channel = endpoint
+                .tls_config(tonic::transport::ClientTlsConfig::new().with_native_roots())
+                .map_err(|err| RemoteProverError::ConnectionFailed(err.into()))?
+                .connect()
                 .await
-                .map_err(|_| RemoteProverError::ConnectionFailed(self.endpoint.to_string()))?
+                .map_err(|err| RemoteProverError::ConnectionFailed(err.into()))?;
+            ApiClient::new(channel)
         };
 
         *client = Some(new_client);